ai-news-summarizer/pyproject.toml
[project]
name = "ai-news-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
